meccec: implement retries in status cmd
Sometimes the mailbox is busy, and becomes impossible to access it and
can cause to hit the timeout.

If this happens during the interrupt handling, it will not reset the
trigger cause, making it impossible to reset it again after.

This implements a retry mechanism to force the status command.